oAuth/03_authorize_endpoints/oauth_service/pyproject.toml
[tool.poetry]
name = "oauth-service"
version = "0.1.0"
description = ""
authors = ["hmzi67 <hamzawaheed057@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
